+++ CHANGELOG.md
@@ -11,6 +11,10 @@ the code was deployed.
 
 ## [Unreleased]
 
+### Added
+
+- Buttons Vitals cache table of only the most recent heartbeat from each Button (CU-2dm6xaf).
+
 ### Removed
 
 - Auto-fresh on Vitals pages.